biondi_client/requirements.txt
PyAutoGUI==0.9.50
Pillow==7.2.0
